security: sanitize Vue directive attributes
